Drop Svelte 2 from HMR
