+++ lib/hot-api.js
@@ -1,51 +1,103 @@
-import { Registry, configure as configureProxy, createProxy } from 'svelte-dev-helper';
+import { makeApplyHmr } from 'svelte-hmr/runtime';
 
-let hotOptions = {
-	noPreserveState: false
+// eslint-disable-next-line no-undef
+const g = typeof window !== 'undefined' ? window : global;
+
+const globalKey =
+	typeof Symbol !== 'undefined'
+		? Symbol('SVELTE_LOADER_HOT')
+		: '__SVELTE_LOADER_HOT';
+
+if (!g[globalKey]) {
+	// do updating refs counting to know when a full update has been applied
+	let updatingCount = 0;
+
+	const notifyStart = () => {
+		updatingCount++;
+	};
+
+	const notifyError = reload => err => {
+		const errString = (err && err.stack) || err;
+		// eslint-disable-next-line no-console
+		console.error(
+			'[HMR] Failed to accept update (nollup compat mode)',
+			errString
+		);
+		reload();
+		notifyEnd();
+	};
+
+	const notifyEnd = () => {
+		updatingCount--;
+		if (updatingCount === 0) {
+			// NOTE this message is important for timing in tests
+			// eslint-disable-next-line no-console
+			console.log('[HMR:Svelte] Up to date');
+		}
+	};
+
+	g[globalKey] = {
+		hotStates: {},
+		notifyStart,
+		notifyError,
+		notifyEnd,
+	};
+}
+
+const runAcceptHandlers = acceptHandlers => {
+	const queue = [...acceptHandlers];
+	const next = () => {
+		const cur = queue.shift();
+		if (cur) {
+			return cur(null).then(next);
+		} else {
+			return Promise.resolve(null);
+		}
+	};
+	return next();
 };
 
-export function configure(options) {
-	hotOptions = Object.assign(hotOptions, options);
-	configureProxy(hotOptions);
-}
+export const applyHmr = makeApplyHmr(args => {
+	const { notifyStart, notifyError, notifyEnd } = g[globalKey];
+	const { m, reload } = args;
 
-export function register(id, component) {
+	let acceptHandlers = (m.hot.data && m.hot.data.acceptHandlers) || [];
+	let nextAcceptHandlers = [];
 
-	//store original component in registry
-	Registry.set(id, {
-		rollback: null,
-		component,
-		instances: []
+	m.hot.dispose(data => {
+		data.acceptHandlers = nextAcceptHandlers;
 	});
 
-	//create the proxy itself
-	const proxy = createProxy(id);
+	const dispose = (...args) => m.hot.dispose(...args);
 
-	//patch the registry record with proxy constructor
-	const record = Registry.get(id);
-	record.proxy = proxy;
-	Registry.set(id, record);
+	const accept = handler => {
+		if (nextAcceptHandlers.length === 0) {
+			m.hot.accept();
+		}
+		nextAcceptHandlers.push(handler);
+	};
 
-	return proxy;
-}
+	const check = status => {
+		if (status === 'ready') {
+			notifyStart();
+		} else if (status === 'idle') {
+			runAcceptHandlers(acceptHandlers)
+				.then(notifyEnd)
+				.catch(notifyError(reload));
+		}
+	};
 
-export function reload(id, component) {
+	m.hot.addStatusHandler(check);
 
-	const record = Registry.get(id);
-
-	//keep reference to previous version to enable rollback
-	record.rollback = record.component;
-
-	//replace component in registry with newly loaded component
-	record.component = component;
-
-	Registry.set(id, record);
-
-	//re-render the proxy instances
-	record.instances.slice().forEach(function(instance) {
-		instance && instance._rerender();
+	m.hot.dispose(() => {
+		m.hot.removeStatusHandler(check);
 	});
 
-	//return the original proxy constructor that was `register()`-ed
-	return record.proxy;
-}
+	const hot = {
+		data: m.hot.data,
+		dispose,
+		accept,
+	};
+
+	return Object.assign({}, args, { hot });
+});
