+++ lib/resolve-svelte.js
@@ -25,16 +25,9 @@ const resolveSvelte = () => {
 };
 
 const { req, base } = resolveSvelte();
-
-const { version } = req(`${base}/package.json`);
-
-const major_version = +version[0];
-
-const { compile, preprocess, walk } =
-  major_version >= 3 ? req(`${base}/compiler`) : req(`${base}`);
+const { compile, preprocess, walk } = req(`${base}/compiler`);
 
 module.exports = {
-	major_version,
 	compile,
 	preprocess,
 	walk,
@@ -42,5 +35,4 @@ module.exports = {
 
 // NOTE svelte3/make-hot requires this module to get the walk function, so we
 // need to have walk available on module.exports before requiring it from here
-module.exports.makeHot =
-  major_version >= 3 ? require('./svelte3/make-hot') : require('./make-hot');
+module.exports.makeHot = require('./make-hot');
