--- lib/svelte3/make-hot.js
@@ -1,12 +1,0 @@
-const { walk } = require('../resolve-svelte');
-const { createMakeHot } = require('svelte-hmr');
-
-const hotApi = require.resolve('./hot-api.js');
-
-const makeHot = createMakeHot({
-	walk,
-	meta: 'module',
-	hotApi,
-});
-
-module.exports = makeHot;
